Create new Robot.cpp, for the 2020 FRC MentorBot.
This is an updated copy of the MentorBot.cpp written for the 2019 FRC tools.  This version builds with the 2020 tools, and has been verified to run correctly on the MentorBot chassis.  Changes were to adjust the include files slightly for the new 2020 style requested by wpilib, and to reassign CANbus numbers to fit where the actual devices were located on the chassis (I had to reverse m_motorLSMaster and m_motorRSMaster).