Melanie's changes to backup if too close to target.
Melanie's changes to use the "Y" (height) limelight value,
rather than the "A" (area) value, to determine how far the
robot is from the target reflective tape -- and added a clause
to back up if too close.

Added code to set limits and ramp values for the drive motors, to
display the motor speeds using Talon-connected magnetic encoders, and
to add a new section to control the drive motors separately if the
joystick trigger is pressed, to use when testing another set of motors
in a test mockup of a 2-motor powercell ball launcher.